refector : 자동차경주 오류 수정
