+++ src/main/java/step3_racingcar/README.md
@@ -12,5 +12,3 @@
 * 전진하는 조건은 0에서 9 사이에서 random 값을 구한 후 random 값이 4이상일 경우이다.
 * 자동차의 상태를 화면에 출력한다. 어느 시점에 출력할 것인지에 대한 제약은 없다.
 
-테스트케이스
------
