+++ src/main/java/step3_racingcar/SimpleRacingCarGame.java
@@ -1,13 +1,13 @@
 package step3_racingcar;
 
 import step3_racingcar.domain.Cars;
+import step3_racingcar.service.GameProcess;
 import step3_racingcar.view.InputView;
 import step3_racingcar.view.ResultView;
 
 public class SimpleRacingCarGame {
 
 	public static void main(String[] args) {
-
 		int carCount = InputView.getCarCount();
 		int tryCount = InputView.getTryCount();
